refactor login page to use oauth hooks
